Feat: Updated Crud for Plugins
Plugin creation now stores the source code and creates logs for it. The
detail view shows more details on the actual Plugin and Plugin Source.

When the PluginSource is added successfuly, the Plugin is extracted and
installed in its own virtual env.

Can now delete a Plugin which will also create logs to delete the
PluginSource, however the source is still kept in the system for
forensic purposes along with all the logs.