Feat: Added plugin run functionality
Implemented crontab functionality with Kronos. Kronos registers the
method in the django project and enable the django settings.

Crontabs are set to run every minute but will only run the plugin if the
interval of that plugin is elapsed.

Plugins are executed on separate threads so that multiple threads can
run at the same time.