Feat: Refactored DB and Subprocess logic
Refact: DB will now soft delete objects instead of deleting them from the system. This will allow deletions to be tracked.

Feat: Subprocesses are now run on separate threads so as not to block the core system.

Refact: Changed datetime.now() to timezone.now() instead of making the datetime date aware of the timezone.